Getting started with TCP. Ithakicopter finished
In the current moment the for loop in the ithakicopter TCP isn't working as expected because probably when using a socket with tcp you are establishing a one-one request-response and it is different than the UDP mindset, where you send a package and the server blasts with responses hoping to get them right. So if you want to receive another tcp response, you need to create another request, another tcp socket in other ways. The read method of the inputstream will work only once per socket as I understood