Create a user interface. Autopilot now is working
In the send section of the autopilot we initiated the tcp socket before sent. That way we don't get a broken pipe each time we are going to need to send something via TCP. I don't understand why though. Also the output stream of the server doesn't contain only telemetry data