Reimplement the tcp telemetry designing approach. Autopilot working fine
For the TCP we have found that reading from the InputStream until null isn't a good approach because it is very time consuming and the null indicated end of the stream. So we needed to know how much bytes are actually sent and to fetch those without expecting null. In other words the time consuming part wasn't to create new tcp sockets but to expect a null to read a data. Thus in order to expect a fixed amount of bytes, each tcp function re-creates a tcp socket, parses the info and get the desired data with a snappy way. Can't find disadvantage of recreating socket instead of using one and trying to handle the streams